+++ apps/decodex-app/Sources/DecodexApp/AccountRunStripScrolling.swift
@@ -84,210 +84,11 @@ final class AccountRunStripScrollProxy {
 	}
 }
 
-private final class AccountRunStripNSScrollView: NSScrollView {
-	var onScrollWheelEvent: ((NSEvent) -> Bool)?
-
-	override func scrollWheel(with event: NSEvent) {
-		if onScrollWheelEvent?(event) == true {
-			return
-		}
-
-		super.scrollWheel(with: event)
-	}
-}
-
-private final class AccountRunStripClipView: NSClipView {
-	var onBoundsChange: (() -> Void)?
-
-	override func constrainBoundsRect(_ proposedBounds: NSRect) -> NSRect {
-		var constrainedBounds = super.constrainBoundsRect(proposedBounds)
-		constrainedBounds.origin.x = max(0, constrainedBounds.origin.x)
-		constrainedBounds.origin.y = max(0, constrainedBounds.origin.y)
-
-		return constrainedBounds
-	}
-
-	override func scroll(to newOrigin: NSPoint) {
-		let oldOrigin = bounds.origin
-
-		super.scroll(to: newOrigin)
-		publishIfNeeded(from: oldOrigin)
-	}
-
-	override func setBoundsOrigin(_ newOrigin: NSPoint) {
-		let oldOrigin = bounds.origin
-
-		super.setBoundsOrigin(newOrigin)
-		publishIfNeeded(from: oldOrigin)
-	}
-
-	private func publishIfNeeded(from oldOrigin: NSPoint) {
-		guard oldOrigin != bounds.origin else {
-			return
-		}
-
-		onBoundsChange?()
-	}
-}
-
-private final class AccountRunDragHostingView<Content: View>: NSHostingView<Content> {
-	weak var dragScrollView: NSScrollView?
-	var onDragScroll: (() -> Void)?
-	var onClick: ((NSPoint) -> Void)?
-	private var dragStartPoint: NSPoint?
-	private var dragStartOffset: CGFloat = 0
-	private var isDraggingContent = false
-
-	override func resetCursorRects() {
-		super.resetCursorRects()
-
-		addCursorRect(bounds, cursor: canDrag ? .openHand : .arrow)
-	}
-
-	override func mouseDown(with event: NSEvent) {
-		guard canDrag, let dragScrollView else {
-			super.mouseDown(with: event)
-			return
-		}
-
-		dragStartPoint = convert(event.locationInWindow, from: nil)
-		dragStartOffset = dragScrollView.contentView.bounds.origin.x
-		isDraggingContent = false
-		NSCursor.openHand.set()
-	}
-
-	override func mouseDragged(with event: NSEvent) {
-		guard
-			let dragStartPoint,
-			let dragScrollView,
-			canDrag
-		else {
-			super.mouseDragged(with: event)
-			return
-		}
-
-		let currentPoint = convert(event.locationInWindow, from: nil)
-		let deltaX = currentPoint.x - dragStartPoint.x
-		if abs(deltaX) > AccountRunStripLayout.dragActivationDistance {
-			isDraggingContent = true
-		}
-		guard isDraggingContent else {
-			return
-		}
-
-		NSCursor.closedHand.set()
-		scroll(dragScrollView, to: dragStartOffset - deltaX)
-	}
-
-	override func mouseUp(with event: NSEvent) {
-		guard canDrag else {
-			super.mouseUp(with: event)
-			return
-		}
-
-		if isDraggingContent == false {
-			onClick?(convert(event.locationInWindow, from: nil))
-		}
-
-		dragStartPoint = nil
-		isDraggingContent = false
-		NSCursor.openHand.set()
-	}
-
-	private var canDrag: Bool {
-		guard let dragScrollView else {
-			return false
-		}
-
-		let contentWidth = dragScrollView.documentView?.frame.width ?? 0
-		return contentWidth > dragScrollView.contentView.bounds.width + 1
-	}
-
-	private func scroll(_ scrollView: NSScrollView, to offset: CGFloat) {
-		let clipView = scrollView.contentView
-		let contentWidth = scrollView.documentView?.frame.width ?? 0
-		let maxOffset = max(0, contentWidth - clipView.bounds.width)
-		let clampedOffset = min(max(0, offset), maxOffset)
-
-		clipView.scroll(to: NSPoint(x: clampedOffset, y: clipView.bounds.origin.y))
-		scrollView.reflectScrolledClipView(clipView)
-		onDragScroll?()
-	}
-}
-
-private final class AccountRunContinuousScroller {
-	private var frameAction: ((TimeInterval) -> Bool)?
-	private var lastTickTime: TimeInterval?
-	private var timer: Timer?
-	private var timerTarget: AccountRunContinuousTimerTarget?
-
-	deinit {
-		stop()
-	}
-
-	func start(_ frameAction: @escaping (TimeInterval) -> Bool) {
-		stop()
-		self.frameAction = frameAction
-		lastTickTime = ProcessInfo.processInfo.systemUptime
-
-		let timerTarget = AccountRunContinuousTimerTarget(scroller: self)
-		let timer = Timer(
-			timeInterval: AccountRunStripLayout.continuousScrollTickInterval,
-			target: timerTarget,
-			selector: #selector(AccountRunContinuousTimerTarget.timerDidFire(_:)),
-			userInfo: nil,
-			repeats: true
-		)
-		self.timerTarget = timerTarget
-		self.timer = timer
-		RunLoop.main.add(timer, forMode: .common)
-	}
-
-	func stop() {
-		timer?.invalidate()
-		timer = nil
-		timerTarget = nil
-		frameAction = nil
-		lastTickTime = nil
-	}
-
-	fileprivate func performFrame() {
-		guard let frameAction else {
-			return
-		}
-
-		let tickTime = ProcessInfo.processInfo.systemUptime
-		let elapsedTime = lastTickTime.map { tickTime - $0 }
-			?? AccountRunStripLayout.continuousScrollTickInterval
-		lastTickTime = tickTime
-
-		let boundedElapsedTime = min(
-			max(elapsedTime, 0),
-			AccountRunStripLayout.continuousScrollMaximumFrameInterval
-		)
-		if frameAction(boundedElapsedTime) == false {
-			stop()
-		}
-	}
-}
-
-private final class AccountRunContinuousTimerTarget: NSObject {
-	weak var scroller: AccountRunContinuousScroller?
-
-	init(scroller: AccountRunContinuousScroller) {
-		self.scroller = scroller
-	}
-
-	@objc func timerDidFire(_ timer: Timer) {
-		scroller?.performFrame()
-	}
-}
-
 final class AccountRunStripContainerView<Content: View>: NSView, AccountRunStripScrollable {
-	fileprivate let scrollView = AccountRunStripNSScrollView()
-	fileprivate let notifyingClipView = AccountRunStripClipView()
-	fileprivate let continuousScroller = AccountRunContinuousScroller()
-	fileprivate let hostingView: AccountRunDragHostingView<Content>
+	let scrollView = AccountRunStripNSScrollView()
+	let notifyingClipView = AccountRunStripClipView()
+	let continuousScroller = AccountRunContinuousScroller()
+	let hostingView: AccountRunDragHostingView<Content>
 	let placementStore: AccountRunStripPlacementStore
 	var measuredContentWidth: CGFloat = 0
 	var onMetricsChange: ((AccountRunStripMetrics) -> Void)?
@@ -393,145 +194,5 @@ final class AccountRunStripContainerView<Content: View>: NSView, AccountRunStrip
 
 	func publishMetrics() {
 		onMetricsChange?(currentMetrics)
-	}
-}
-
-extension AccountRunStripContainerView {
-	func scroll(by distance: CGFloat) {
-		scroll(to: clipView.bounds.origin.x + distance)
-	}
-
-	func scrollToAdjacentRun(_ direction: AccountRunStripScrollDirection) {
-		layoutSubtreeIfNeeded()
-
-		guard let offset = adjacentRunOffset(for: direction) else {
-			return
-		}
-
-		scroll(to: offset, animated: true)
-	}
-
-	func startContinuousScroll(_ direction: AccountRunStripScrollDirection) {
-		layoutSubtreeIfNeeded()
-
-		guard measuredContentWidth > clipView.bounds.width + 1 else {
-			return
-		}
-
-		continuousScroller.start { [weak self] elapsedTime in
-			guard let self else {
-				return false
-			}
-
-			let previousOffset = self.clipView.bounds.origin.x
-			let distance = direction.scrollMultiplier
-				* AccountRunStripLayout.continuousScrollVelocity
-				* CGFloat(elapsedTime)
-			self.scroll(by: distance)
-
-			return previousOffset != self.clipView.bounds.origin.x
-		}
-	}
-
-	func stopContinuousScroll() {
-		continuousScroller.stop()
-	}
-
-	func handleScrollWheel(_ event: NSEvent) -> Bool {
-		layoutSubtreeIfNeeded()
-
-		guard measuredContentWidth > clipView.bounds.width + 1 else {
-			return false
-		}
-
-		let distance = wheelScrollDistance(from: event)
-		guard abs(distance) > AccountRunStripLayout.wheelMinimumDelta else {
-			return false
-		}
-
-		let previousOffset = clipView.bounds.origin.x
-		scroll(by: distance)
-
-		return previousOffset != clipView.bounds.origin.x
-	}
-
-	func scrollClickedRunToLeadingEdge(at point: NSPoint) {
-		layoutSubtreeIfNeeded()
-
-		guard
-			measuredContentWidth > clipView.bounds.width + 1,
-			let runID = placementStore.runID(containing: point),
-			let frame = placementStore.frame(for: runID)
-		else {
-			return
-		}
-
-		scroll(to: frame.minX, animated: true)
-	}
-
-	private func adjacentRunOffset(for direction: AccountRunStripScrollDirection) -> CGFloat? {
-		let maxOffset = max(0, measuredContentWidth - clipView.bounds.width)
-		guard maxOffset > 0 else {
-			return nil
-		}
-
-		let currentOffset = clipView.bounds.origin.x
-		let orderedOffsets = placementStore.orderedFrames().map(\.minX)
-		let targetOffset: CGFloat?
-		switch direction {
-		case .backward:
-			targetOffset = orderedOffsets.last { offset in
-				offset < currentOffset - 1
-			} ?? (currentOffset > 0 ? 0 : nil)
-		case .forward:
-			targetOffset = orderedOffsets.first { offset in
-				offset > currentOffset + 1
-			} ?? (currentOffset < maxOffset ? maxOffset : nil)
-		}
-
-		return targetOffset.map { min(max(0, $0), maxOffset) }
-	}
-
-	private func wheelScrollDistance(from event: NSEvent) -> CGFloat {
-		let rawDeltaX = event.scrollingDeltaX == 0 ? event.deltaX : event.scrollingDeltaX
-		let rawDeltaY = event.scrollingDeltaY == 0 ? event.deltaY : event.scrollingDeltaY
-		let deltaX = scaledWheelDelta(rawDeltaX, isPrecise: event.hasPreciseScrollingDeltas)
-		let deltaY = scaledWheelDelta(rawDeltaY, isPrecise: event.hasPreciseScrollingDeltas)
-		let dominantDelta = abs(deltaX) >= abs(deltaY) ? deltaX : deltaY
-
-		return -dominantDelta
-	}
-
-	private func scaledWheelDelta(_ delta: CGFloat, isPrecise: Bool) -> CGFloat {
-		isPrecise ? delta : delta * AccountRunStripLayout.wheelLineDeltaScale
-	}
-
-	func scroll(to offset: CGFloat, animated: Bool = false) {
-		layoutSubtreeIfNeeded()
-
-		let maxOffset = max(0, measuredContentWidth - clipView.bounds.width)
-		let clampedOffset = min(max(0, offset), maxOffset)
-		guard clampedOffset != clipView.bounds.origin.x else {
-			return
-		}
-
-		if animated {
-			animateScroll(to: clampedOffset)
-			return
-		}
-
-		clipView.scroll(to: NSPoint(x: clampedOffset, y: clipView.bounds.origin.y))
-		scrollView.reflectScrolledClipView(clipView)
-		publishMetrics()
-	}
-
-	private func animateScroll(to offset: CGFloat) {
-		NSAnimationContext.runAnimationGroup { context in
-			context.duration = AccountRunStripLayout.clickScrollDuration
-			context.allowsImplicitAnimation = true
-			clipView.animator().setBoundsOrigin(NSPoint(x: offset, y: clipView.bounds.origin.y))
-		}
-		scrollView.reflectScrolledClipView(clipView)
-		publishMetrics()
 	}
 }
